[Feat] #14 - type 별로 오른쪽 버튼 설정
